html/browsers/browsing-the-web/back-forward-cache/dedicated-worker-broadcastchanel.html

<!doctype html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
promise_test(async t => {
  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;

  window.open(urlA, '_blank', 'noopener');
  await pageA.execute_script(waitForPageShow);

  await pageA.execute_script(() => {
    // Create BroadcastChannel in main thread
    let channel = new BroadcastChannel('foo');
    // Register a dedicated worker.
    let worker = new Worker('../resources/worker-with-broadcastchannel.js');
    channel.addEventListener('message', event => {
        // Received from worker: event.data;
    });
    worker.postMessage({channelName: 'foo'});
  });

  await navigateAndThenBack(pageA, pageB, urlB);
  await assert_bfcached(pageA);

}, 'PageA should be restored from bfcache, even though the dedicated worker had an open broadcastchannel connection.');
</script>
